pyproject.toml
[tool.poetry]
name = "nejemojo"
version = "0.1.0"
description = "Drives your NEJE desktop engraver"
authors = ["sthysel <sthysel@gmail>"]
license = "GPL"

packages = [
    { include = "nejemojo", from = "src" }
]

[tool.poetry.scripts]
nejemojo = "nejemojo.cli:neje"
nejeview = "nejemojo.cli:view"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.3.1"
pyserial = "^3.5"
click = "^8.0.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
